refactor: доработана предыдущая логика, хендлеры исключений вынесены в отдельный класс, переработана логика моделей, убраны директории checked/unchecked исключений
