+++ src/main/java/ru/yandex/practicum/filmorate/exceptions/ObjectAlreadyExistException.java
@@ -1,4 +1,4 @@
-package ru.yandex.practicum.filmorate.exceptions.unchecked;
+package ru.yandex.practicum.filmorate.exceptions;
 
 public class ObjectAlreadyExistException extends RuntimeException{
     public ObjectAlreadyExistException() {
