+++ src/main/java/ru/yandex/practicum/filmorate/exceptions/ObjectNotFoundException.java
@@ -1,4 +1,4 @@
-package ru.yandex.practicum.filmorate.exceptions.unchecked;
+package ru.yandex.practicum.filmorate.exceptions;
 
 public class ObjectNotFoundException extends RuntimeException {
     public ObjectNotFoundException() {
