+++ src/main/java/ru/yandex/practicum/filmorate/exceptions/UniqueObjectException.java
@@ -1,4 +1,4 @@
-package ru.yandex.practicum.filmorate.exceptions.unchecked;
+package ru.yandex.practicum.filmorate.exceptions;
 
 public class UniqueObjectException extends RuntimeException {
     public UniqueObjectException() {
