+++ src/main/java/ru/yandex/practicum/filmorate/exceptions/ValidateException.java
@@ -1,4 +1,4 @@
-package ru.yandex.practicum.filmorate.exceptions.unchecked;
+package ru.yandex.practicum.filmorate.exceptions;
 
 public class ValidateException extends RuntimeException {
     public ValidateException(String msg) {
